add tx_status for marketplace builder

+++ legacy/src/builder_state.rs
@@ -677,7 +677,7 @@ impl<Types: NodeType> BuilderState<Types> {
                 *tx,
                 TransactionStatus::Sequenced {
                     block: self.parent_block_references.view_number.u64(),
-                    offset: 0 as u64,
+                    offset: 0_u64,
                 },
             );
         }
